+++ PreflightApi.API/Controllers/RunwayController.cs
@@ -106,8 +106,8 @@ public class RunwayController(IRunwayService runwayService) : ControllerBase
     [ProducesResponseType(typeof(PaginatedResponse<RunwayDto>), StatusCodes.Status200OK)]
     [ProducesResponseType(typeof(ApiErrorResponse), StatusCodes.Status400BadRequest)]
     public async Task<ActionResult<PaginatedResponse<RunwayDto>>> SearchNearby(
-        [FromQuery] decimal lat,
-        [FromQuery] decimal lon,
+        [FromQuery] double lat,
+        [FromQuery] double lon,
         [FromQuery] PaginationParams pagination,
         CancellationToken ct,
         [FromQuery] double radiusNm = 30,
