+++ PreflightApi.Infrastructure/Interfaces/IAirportService.cs
@@ -8,5 +8,5 @@ public interface IAirportService
     Task<PaginatedResponse<AirportDto>> GetAirports(string? search, string[]? stateCodes, string? cursor, int limit, CancellationToken ct = default);
     Task<AirportDto> GetAirportByIcaoCodeOrIdent(string icaoCodeOrIdent, CancellationToken ct = default);
     Task<IEnumerable<AirportDto>> GetAirportsByIcaoCodesOrIdents(string[] codesOrIdents, CancellationToken ct = default);
-    Task<PaginatedResponse<AirportDto>> SearchNearby(decimal latitude, decimal longitude, double radiusNm, string? cursor, int limit, CancellationToken ct = default);
+    Task<PaginatedResponse<AirportDto>> SearchNearby(double latitude, double longitude, double radiusNm, string? cursor, int limit, CancellationToken ct = default);
 }
